requirements.txt
# backend/requirements.txt
# Flaskと関連ライブラリ
Flask==3.1.2
flask-cors==4.0.0
gunicorn==21.2.0

# 環境変数管理
python-dotenv==1.0.0

# データ処理
pandas==2.1.0
numpy==1.24.3

# Google Sheets連携
gspread==5.7.2
oauth2client==4.1.3

# HTTPリクエスト
requests==2.31.0

# セキュリティ
cryptography==41.0.3

# 開発用ツール
pytest==7.4.0
coverage==7.3.0

# ロギング
logging==0.5.1.2

# Google API
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.100.0
